Relabeled the command "New From Template" to "New File from Template"
In order to improve readability and clarity, changed the label for the
command to create a new class/test/interface/exception etc.
"New from Template" --> "New File from Template"

Also updated the label references in READMEs, messages and translations.

This also avoids the command "New File" getting repeated twice in the
Project explorer panel if it is simplified further as done in PR #302.

Signed-off-by: Siddharth Srinivasan <siddharth.srinivasan@oracle.com>