Fix NetBeans launch failure on unix shell for arguments with quotes et al
Backporting Netbeans PR #7908 (still under review) in order for the LSP
server launch to work for launch arguments including the project path
containing special characters like quotes, spaces, backticks and $.

This is an interim patch to mitigate the effect on user experience,
while the upstream PR is under review. It will be updated to the
approved patch, as and when it is made available.

This is partially related to #301, which only fixes for spaces in
arguments.

Additionally, the same diff is applied to script/bin/nbcode which is a
copy of the netbeans java.lsp.server module's copy.

Signed-off-by: Siddharth Srinivasan <siddharth.srinivasan@oracle.com>